Back-end/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==4.2.29
gunicorn==23.0.0
idna==3.11
PyMySQL==1.1.2
pypdf==6.1.1
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.8.2
cryptography>=41.0.0
django-redis==6.0.0
redis==7.3.0
Pillow
djangorestframework==3.17.0
django-storages==1.14.4
boto3==1.35.99
beautifulsoup4==4.12.3
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
discord.py==2.5.2
pytest==9.0.3
pytest-django==4.12.0
holidays>=0.40,<1.0
